Add files via upload
Prim's algorithm to find minimum cost spanning tree (as Kruskal's algorithm) uses the greedy approach. Prim's algorithm shares a similarity with the shortest path first algorithms.
Prim’s Algorithm also use Greedy approach to find the minimum spanning tree. In Prim’s Algorithm we grow the spanning tree from a starting position. Unlike an edge in Kruskal's, we add vertex to the growing spanning tree in Prim's.

The implementation discussed in Prims minimum cost uses two arrays to find minimum weight edge that connects the two sets. Here we use one in MST [V]. The value of MST[i] is going to be true if vertex i is included in the MST. In every pass, we consider only those edges such that one vertex of the edge is included in MST and other is not. After we pick an edge, we mark both vertices as included in MST.

Time Complexity : O (V3)

Note that time complexity of previous approach that uses adjacency matrix is O(V2) and time complexity of the adjacency list representation implementation is O((E+V)LogV.